refactor(server): extract ACP boot into boot-acp.ts (#4243 step 2)
